캠페인 생성시 ONOFF기본값을 ON, 마케터 actionlog 활성화, 라우트 파일 정리
